feat. 카드 교환 로직 개발 (단순 AccountCard 엔티티의 데이터를 조작하는 코드)
추후 교환 엔티티 만들고 해당 엔티티를 통한 교환 로직으로 변경해야 함.

+++ src/main/java/chimhaha/chimcard/card/dto/MyCardResponseDto.java
@@ -12,9 +12,9 @@ public class MyCardResponseDto {
     private final Grade grade;
     private final int power;
     private final String cardSeason;
-    private final int count;
+    private final long count;
 
-    public MyCardResponseDto(Card card, int count) {
+    public MyCardResponseDto(Card card, long count) {
         this.id = card.getId();
         this.title = card.getTitle();
         this.attackType = card.getAttackType().getType();
